+++ src/QueueConnection.js
@@ -18,7 +18,7 @@ class QueueConnection extends EventEmitter {
     this._connectionPromise = null
     this._channel = null
     this._channelPromise = null
-    this._activeConnectionConfig = null
+    this._activeConnectionUrl = null
   }
 
   setLogger (logger) {
@@ -47,7 +47,8 @@ class QueueConnection extends EventEmitter {
     }
 
     this._connectionPromise = this._connect(this._config.url, options).then((conn) => {
-      this._logger.info(`RabbitMQ connection established on '${this._config.url.hostname}' host`)
+      const urlObject = new URL(this._activeConnectionUrl)
+      this._logger.info(`RabbitMQ connection established on '${urlObject.host}' host`)
 
       conn.on('error', (err) => {
         if (err.message !== 'Connection closing') {
@@ -85,36 +86,21 @@ class QueueConnection extends EventEmitter {
       for (const url of configUrl) {
         try {
           const connection = await amqp.connect(url, options)
-          this._activeConnectionConfig = url
+          this._activeConnectionUrl = url
           return connection
         } catch (err) {
           // let the next connection url in the list by tried
-          this._logger.warn('RabbitMQ connection failed to url:', url)
+          const urlObject = new URL(url)
+          this._logger.warn('RabbitMQ connection failed to host:', urlObject.host)
         }
       }
 
       throw new Error('RabbitMQ connection filed with multiple urls')
-    } else if (Array.isArray(configUrl.hostname)) {
-      // handle multiple connection hosts
-      for (const host of configUrl.hostname) {
-        const connectionUrl = {
-          ...this._config.url, // copy given config
-          hostname: host // use hostname from current iteration
-        }
-
-        try {
-          const connection = await amqp.connect(connectionUrl, options)
-          this._activeConnectionConfig = connectionUrl
-          return connection
-        } catch (err) {
-          this._logger.warn('RabbitMQ connection failed to host:', { ...connectionUrl, password: connectionUrl.password ? '***' : connectionUrl.password })
-        }
-      }
-
-      throw new Error('RabbitMQ connection filed with multiple hosts')
     } else {
-      // assume simple url string or standard url object
-      return amqp.connect(configUrl, options)
+      // assume simple url string
+      const connection = await amqp.connect(configUrl, options)
+      this._activeConnectionUrl = configUrl
+      return connection
     }
   }