+++ test/QueueConnection.test.js
@@ -45,7 +45,7 @@ describe('QueueConnection', () => {
 
     const connection = new QueueConnection(multiUrlConfig)
     await connection.connect()
-    assert.strictEqual(connection._activeConnectionConfig, config.url)
+    assert.strictEqual(connection._activeConnectionUrl, config.url)
   })
 
   it('#connect() handles multiple string urls and tries the next url in the list if one is not working', async () => {
@@ -55,43 +55,7 @@ describe('QueueConnection', () => {
 
     const connection = new QueueConnection(multiUrlConfig)
     await connection.connect()
-    assert.strictEqual(connection._activeConnectionConfig, config.url)
-  })
-
-  it('#connect() handles multiple hosts in an url object and connects to the first working one', async () => {
-    const url = new URL(config.url)
-    const urlObject = {
-      protocol: url.protocol.replace(':', ''),
-      hostname: [url.hostname, 'invalid_host'],
-      port: parseInt(url.port, 10),
-      username: url.username ? url.username : undefined,
-      password: url.password ? url.password : undefined
-    }
-    const multiUrlConfig = copyConfig({
-      url: urlObject
-    })
-
-    const connection = new QueueConnection(multiUrlConfig)
-    await connection.connect()
-    assert.strictEqual(connection._activeConnectionConfig.hostname, url.hostname)
-  })
-
-  it('#connect() handles multiple hosts in an url object and tries the next url in the list if one is not working', async () => {
-    const url = new URL(config.url)
-    const urlObject = {
-      protocol: url.protocol.replace(':', ''),
-      hostname: [url.hostname, 'invalid_host'],
-      port: parseInt(url.port, 10),
-      username: url.username ? url.username : undefined,
-      password: url.password ? url.password : undefined
-    }
-    const multiUrlConfig = copyConfig({
-      url: urlObject
-    })
-
-    const connection = new QueueConnection(multiUrlConfig)
-    await connection.connect()
-    assert.strictEqual(connection._activeConnectionConfig.hostname, url.hostname)
+    assert.strictEqual(connection._activeConnectionUrl, config.url)
   })
 
   it('#close() closes connection to RabbitMQ', async () => {
@@ -127,7 +91,7 @@ describe('QueueConnection', () => {
     }
 
     try {
-      await connection.close(true)
+      await connection.close()
     } catch (e) {
       throw new Error(`close() failed: ${e}`)
     }